pyproject.toml
[project]
name = "g-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "camelot-py[cv]>=1.0.0",
    "faiss-gpu",
    "ffmpeg-python>=0.2.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.24",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.4",
    "langchain-neo4j>=0.4.0",
    "langchain-openai>=0.3.15",
    "moviepy>=1.0.3",
    "neo4j>=5.28.1",
    "openai-whisper>=20240930",
    "opencv-python>=4.11.0.86",
    "pdf2docx>=0.5.8",
    "pdfplumber>=0.11.6",
    "pillow>=11.2.1",
    "pydub>=0.25.1",
    "pymupdf>=1.25.5",
    "pymupdf4llm>=0.0.22",
    "pymupdfpro>=1.25.5",
    "pytesseract>=0.3.13",
    "pytube>=15.0.0",
    "sentence-transformers>=4.1.0",
    "speechrecognition>=3.14.2",
    "transformers>=4.51.3",
    "yt-dlp>=2025.4.30",
]
